+++ pkg/resourcemanager/rediscaches/actions/rediscacheaction_reconcile.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-package rediscaches
+package actions
 
 import (
 	"context"
